requirements.txt
mysql_connector_repackaged==0.3.1
PyYAML==6.0
tweepy==4.5.0
